Test: 샘플 데이터 저장
음식점 정보, 방문 정보 저장
